Uniform Integration of Spyder's Run Configuration and PYTHONPATH
Line profiler behavior is the same if started from the run menu or the widget gui. Run configuration is loaded in the widget. File names are normalized with osp.abspath for parity with file name strings from the editor. File names are only added once to the widget's QComboBox. Additional checks are made to properly select the working directory when automatically importing from the run configuration. Spyder's PYTHONPATH is automatically imported from the MainWindow instance passed to the widget.